pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Test project - example of book store"
authors = ["IvanB <baklanovivan25@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-decouple = "^3.8"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "^8.0.0" }
redis = { extras = ["hiredis"], version = "^5.2.1" }
django-debug-toolbar = "^4.4.6"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
pytest-django = "^4.9.0"
mixer = "^7.2.2"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-results = "^2.5.1"
django-redis = "^5.4.0"
django-celery-beat = "^2.7.0"
django-simple-captcha = "^0.6.1"
django-axes = "^7.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests/test_*.py", "tests/*_test.py"]
